% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/olleyes_files_converter.R
\name{convert_pngs_to_pdf}
\alias{convert_pngs_to_pdf}
\title{Convert PNGs to PDF and combine OD and OS PDFs}
\usage{
convert_pngs_to_pdf(file_source, output_file)
}
\arguments{
\item{file_source}{The directory containing PNG files.}

\item{Output_file}{The output PDF file name without .pdf extension.}
}
\description{
Convert PNGs to PDF and combine OD and OS PDFs
}
